Preparation for CrawlyUI integration
