+++ lib/crawly/utils.ex
@@ -139,6 +139,31 @@ defmodule Crawly.Utils do
     end
   end
 
+
+  @doc """
+  Returns a list of known modules which implements Crawly.Spider behaviour
+  """
+  @spec list_spiders() :: [module()]
+  def list_spiders() do
+    Enum.reduce(
+      get_modules_from_applications(),
+      [],
+      fn mod, acc ->
+        try do
+          beh = Keyword.get(mod.__info__(:attributes), :behaviour)
+
+          case beh == [Crawly.Spider] do
+            true ->
+              [mod] ++ acc
+            false ->
+              acc
+          end
+
+        rescue _ -> acc end
+      end)
+  end
+
+
   ##############################################################################
   # Private functions
   ##############################################################################
@@ -158,5 +183,17 @@ defmodule Crawly.Utils do
       false ->
         nil
     end
+  end
+
+  @spec get_modules_from_applications() :: [module()]
+  def get_modules_from_applications do
+    Enum.reduce(Application.started_applications(), [], fn {app, _descr, _vsn}, acc ->
+      case :application.get_key(app, :modules) do
+        {:ok, modules} ->
+          modules ++ acc
+        _other ->
+          acc
+      end
+    end)
   end
 end
